完成了数据库存储读取的功能

+++ OOPDataManage/GeomertyFile/CPoint.cpp
@@ -5,7 +5,7 @@ GeometryType CPoint::GetType()const
 	return GeometryType::Point;
 }
 
-CPoint::CPoint(float x, float y):m_X(x),m_Y(y)
+CPoint::CPoint(double x, double y):m_X(x),m_Y(y)
 {
 	ID = GetNewID(this);
 }
@@ -58,12 +58,12 @@ CPoint& CPoint::operator-=(CPoint c)
 	return *this;
 }
 
-float CPoint::operator*(CPoint c)const
+double CPoint::operator*(CPoint c)const
 {
 	return this->m_X * c.m_X + this->m_Y * c.m_Y;
 }
 
-float CPoint::operator^(CPoint c)const
+double CPoint::operator^(CPoint c)const
 {
 	return this->m_X * c.m_Y - this->m_Y * c.m_X;
 }
@@ -84,10 +84,10 @@ bool CPoint::operator!=(CPoint c)const
 		return true;
 }
 
-float CPoint_Distance(const CPoint& c1, const CPoint& c2)
+double CPoint_Distance(const CPoint& c1, const CPoint& c2)
 {
-	float dx = c1.x() - c2.x();
-	float dy = c1.y()- c2.y();
+	double dx = c1.x() - c2.x();
+	double dy = c1.y()- c2.y();
 	return sqrt(dx*dx+dy*dy);
 }
