+++ OOPDataManage/GeomertyFile/CPoint.h
@@ -9,7 +9,7 @@ class CPoint :public CGeometry,public IDManager<CPoint>
 {
 public:
 	//构造函数
-	CPoint(float x, float y);
+	CPoint(double x, double y);
 	//默认构造函数
 	CPoint();
 	//拷贝构造函数
@@ -20,14 +20,14 @@ public:
 	~CPoint();
 
 	//获取x
-	float x()const { return m_X; };
+	double x()const { return m_X; };
 	//获取y
-	float y()const { return m_Y; };
+	double y()const { return m_Y; };
 
 	//设置x
-	void x(float x) { m_X = x; };
+	void x(double x) { m_X = x; };
 	//设置y
-	void y(float y) { m_Y = y; };
+	void y(double y) { m_Y = y; };
 
 	GeometryType GetType()const override;
 	const char* ToWKT()const override;
@@ -42,9 +42,9 @@ public:
 	//重载-运算
 	CPoint operator - (CPoint)const;
 	//重载*点积运算
-	float operator*(CPoint c)const;
+	double operator*(CPoint c)const;
 	//重载^叉积运算
-	float operator ^ (CPoint c)const;
+	double operator ^ (CPoint c)const;
 	//判断点X、Y值是否相等
 	bool operator==(const CPoint& other) const;
 	//判断点X、Y值是否不等
@@ -52,12 +52,12 @@ public:
 
 protected:
 	//x坐标
-	float m_X;
+	double m_X;
 	//y坐标
-	float m_Y;
+	double m_Y;
 
 };
 
 //求两个点之间的距离
-float CPoint_Distance(const CPoint& c1, const CPoint& c2);
+double CPoint_Distance(const CPoint& c1, const CPoint& c2);
